move AddressBook out of category.mjs
